FossilOrigin-Name: fdd3fe21c25a45cc5c02ea3116d91216aedbfe33a4d1f4b5fe3e8da3b02a702f

+++ src/fileedit.c
@@ -619,7 +619,7 @@ static int checkin_mini(CheckinMiniInfo * pCI, int *pRid, Blob * pErr){
             blob_add_cr(&pCI->fileContent);
           }
         }
-        if(blob_size(&pCI->fileContent)!=oldSize){
+        if((int)blob_size(&pCI->fileContent)!=oldSize){
           rehash = 1;
         }
       }
